Formatting fixes
The test.go file is unnecessary as there should only be one main function per-program and it seams to not do anything. .DS_Store should not be committed as it is a local cache file for your Mac. There should be a .gitignore to prevent committing files like .DS_Store files and IDE config files. You should always run “go fmt” on files before committing to make sure all the code “tabs/spaces” is formatted according to the go standard.